codex: address PR review feedback (#26537)
